+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Types/Input.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE UnboxedTuples, MagicHash, RecordWildCards, DerivingStrategies #-}
+{-# LANGUAGE UnboxedTuples, MagicHash, RecordWildCards #-}
 {-|
 Module      : Parsley.Internal.Backend.Machine.Types.Input
 Description : Packaging of offsets and positions.
@@ -11,7 +11,13 @@ for static augmented information, and `Input#` is a raw combination of the two c
 
 @since 1.8.0.0
 -}
-module Parsley.Internal.Backend.Machine.Types.Input (module Parsley.Internal.Backend.Machine.Types.Input) where
+module Parsley.Internal.Backend.Machine.Types.Input (
+    Input(off), Input#(..),
+    mkInput, fromInput, toInput,
+    consume,
+    forcePos, updatePos,
+    chooseInput
+  ) where
 
 import Parsley.Internal.Backend.Machine.InputRep                  (Rep)
 import Parsley.Internal.Backend.Machine.Types.Input.Offset        (Offset(offset), mkOffset, moveOne, moveN)
@@ -25,13 +31,14 @@ import Parsley.Internal.Core.CombinatorAST                        (PosSelector)
 Packages known static information about offsets (via `Offset`) with static information about positions
 (currently unavailable).
 
-@since 1.8.0.0
+@since 2.1.0.0
 -}
 data Input o = Input {
+    -- | The offset contained within the input
     off  :: !(Offset o),
+    -- | The position contained within the input
     pos :: !StaPos
   }
-  deriving stock Show
 
 {-|
 Packages a dynamic offset with a dynamic position.
@@ -43,27 +50,13 @@ data Input# o = Input# {
     pos#  :: !DynPos
   }
 
+{-|
+Constructs an `Input` given a dynamic offset and a static position.
+
+@since 2.1.0.0
+-}
 mkInput :: Code (Rep o) -> (Word, Word) -> Input o
 mkInput off = Input (mkOffset off 0) . fromStaPos
-
-consume :: Code (Rep o) -> Input o -> Input o
-consume offset' input = input {
-    off = moveOne (off input) offset'
-  }
-
-forcePos :: Input o -> PosSelector -> (Code Int -> Input o -> Code r) -> Code r
-forcePos input sel k = force (pos input) sel (\dp sp -> k dp (input { pos = sp }))
-
-updatePos :: Input o -> Code Char -> CharPred -> Input o
-updatePos input c p = input { pos = update (pos input) c p }
-
--- TODO: Documentation: This function can refine the input passed forward during a call based on the known characteristics about the function.
-chooseInput :: InputCharacteristic -> Word -> Input o -> Input# o -> Input o
-chooseInput (AlwaysConsumes n) _ inp  inp#  = inp { off = moveN n (off inp) (off# inp#), pos = fromDynPos (pos# inp#) }
--- Technically, in this case, we know the whole input is unchanged. This essentially ignores the continuation arguments
--- hopefully GHC could optimise this better?
-chooseInput NeverConsumes      _ inp  _inp# = inp -- { off = (off inp) {offset = off# inp# }, pos = pos# inp# }
-chooseInput MayConsume         u _inp inp#  = toInput u inp#
 
 {-|
 Strips away static information, returning the raw dynamic components.
@@ -80,3 +73,46 @@ Given a unique identifier, forms a plainly annotated static combination of posit
 -}
 toInput :: Word -> Input# o -> Input o
 toInput u Input#{..} = Input (mkOffset off# u) (fromDynPos pos#)
+
+{-|
+Register that a character has been consumed on this input, incorporating the new dynamic offset.
+
+@since 2.1.0.0
+-}
+consume :: Code (Rep o) -> Input o -> Input o
+consume offset' input = input {
+    off = moveOne (off input) offset'
+  }
+
+{-|
+Collapse the position stored inside the input applying all updates to it. Once this has been completed,
+the given `PosSelector` will be used to extract one of the line or column and return it to the given
+continuation, along with the updated input post-collapse.
+
+@since 2.1.0.0
+-}
+forcePos :: Input o -> PosSelector -> (Code Int -> Input o -> Code r) -> Code r
+forcePos input sel k = force (pos input) sel (\dp sp -> k dp (input { pos = sp }))
+
+{-|
+Updates the position within the `Input` when a character has been consumed, providing it the
+dynamic character that was produced as well as the static character-predicate that guarded it.
+
+@since 2.1.0.0
+-}
+updatePos :: Input o -> Code Char -> CharPred -> Input o
+updatePos input c p = input { pos = update (pos input) c p }
+
+{-|
+Given knowledge about how input has been consumed through a call boundary, this function can update
+the input using statically acquired knowledge.
+
+@since 2.1.0.0
+-}
+-- TODO: In future, we could adjust InputCharacteristic to provide information about the static behaviours of the positions too...
+chooseInput :: InputCharacteristic -> Word -> Input o -> Input# o -> Input o
+chooseInput (AlwaysConsumes n) _ inp  inp#  = inp { off = moveN n (off inp) (off# inp#), pos = fromDynPos (pos# inp#) }
+-- Technically, in this case, we know the whole input is unchanged. This essentially ignores the continuation arguments
+-- hopefully GHC could optimise this better?
+chooseInput NeverConsumes      _ inp  _inp# = inp -- { off = (off inp) {offset = off# inp# }, pos = pos# inp# }
+chooseInput MayConsume         u _inp inp#  = toInput u inp#